Move models over to MixItUp.Base project
