+++ MixItUp.Base/Model/DeveloperAPIs/ChatMessageDeveloperAPIModel.cs
@@ -1,9 +1,7 @@
-﻿using MixItUp.Base.ViewModel.User;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
 
-namespace MixItUp.Desktop.Services.DeveloperAPI.Models
+namespace MixItUp.Base.Model.DeveloperAPIs
 {
     [DataContract]
     public class ChatMessageDeveloperAPIModel
