+++ MixItUp.Base/Model/DeveloperAPIs/UserCurrencyGiveDeveloperAPIModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.Serialization;
 
-namespace MixItUp.Desktop.Services.DeveloperAPI.Models
+namespace MixItUp.Base.Model.DeveloperAPIs
 {
     [DataContract]
     public class UserCurrencyGiveDeveloperAPIModel
