+++ MixItUp.Base/Model/DeveloperAPIs/UserCurrencyUpdateDeveloperAPIModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.Serialization;
 
-namespace MixItUp.Desktop.Services.DeveloperAPI.Models
+namespace MixItUp.Base.Model.DeveloperAPIs
 {
     [DataContract]
     public class UserCurrencyUpdateDeveloperAPIModel
