+++ MixItUp.Desktop/MixItUp.Desktop.csproj
@@ -194,14 +194,8 @@
     <Compile Include="Services\DeveloperAPI\CommandController.cs" />
     <Compile Include="Services\DeveloperAPI\CurrencyController.cs" />
     <Compile Include="Services\DeveloperAPI\MixerUserController.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\CommandDeveloperAPIModel.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\ChatMessageDeveloperAPIModel.cs" />
     <Compile Include="Services\DeveloperAPI\NoCacheHeader.cs" />
     <Compile Include="Services\DeveloperAPI\UserController.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\UserCurrencyUpdateDeveloperAPIModel.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\UserCurrencyGiveDeveloperAPIModel.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\UserCurrencyDeveloperAPIModel.cs" />
-    <Compile Include="Services\DeveloperAPI\Models\UserDeveloperAPIModel.cs" />
     <Compile Include="Services\DiscordService.cs" />
     <Compile Include="Services\GameWispService.cs" />
     <Compile Include="Services\GawkBoxService.cs" />
@@ -271,6 +265,7 @@
       <Name>obs-websocket-dotnet</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\System.Data.SQLite.Core.1.0.108.0\build\net46\System.Data.SQLite.Core.targets" Condition="Exists('..\packages\System.Data.SQLite.Core.1.0.108.0\build\net46\System.Data.SQLite.Core.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
