+++ MixItUp.Desktop/Services/DeveloperAPI/CurrencyController.cs
@@ -1,6 +1,6 @@
 ﻿using MixItUp.Base;
+using MixItUp.Base.Model.DeveloperAPIs;
 using MixItUp.Base.ViewModel.User;
-using MixItUp.Desktop.Services.DeveloperAPI.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
